Delete Action and view changes
The code accepts an optional parameter that indicates whether the method
was called after a failure to save changes. This parameter is false when
the HttpGet Delete method is called without a previous failure. When it
is called by the HttpPost Delete method in response to a database update
error, the parameter is true and an error message is passed to the view.