Reading related data
Added instructor page.
When Instructor is selected, the user is getting all the intructor's
courses.
When Course is selected, the user is getting all the students in the
course with their grade.

when selectedRow string variable get value "success", the <th> class
becomes "success", and Bootstrap makes the <th> green. :)